feat: 자신이 아니면 댓글 삭제를 못하도록 설정

+++ src/main/java/codesquad/springcafe/reply/ReplyController.java
@@ -28,9 +28,9 @@ public class ReplyController {
     @DeleteMapping("/{replyId}/delete")
     public String deleteReply(@PathVariable Long replyId, HttpSession httpSession) {
         String userId = httpSession.getAttribute("userId").toString();
-//        if (!replyService.isAutehnticated(userId, replyId)) {
-//            return "redirect:/";
-//        }
+        if (!replyService.isAutehnticated(userId, replyId)) {
+            return "redirect:/";
+        }
         replyService.delete(replyId);
         return "redirect:/";
     }
